+++ apps/auth/nextjs-ssr/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @crossmint/server-sdk-next-starter
 
+## 0.4.25
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@1.16.1
+
+## 0.4.24
+
+### Patch Changes
+
+- Updated dependencies [025a5a2]
+  - @crossmint/client-sdk-react-ui@1.16.0
+
 ## 0.4.23
 
 ### Patch Changes