+++ apps/payments/create-react-app/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @crossmint/client-sdk-react-ui-starter
 
+## 1.3.51
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@1.16.1
+
+## 1.3.50
+
+### Patch Changes
+
+- Updated dependencies [025a5a2]
+  - @crossmint/client-sdk-react-ui@1.16.0
+
 ## 1.3.49
 
 ### Patch Changes