+++ apps/wallets/smart-wallet/next/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @crossmint/client-sdk-smart-wallet-next-starter
 
+## 0.1.79
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@1.16.1
+
+## 0.1.78
+
+### Patch Changes
+
+- Updated dependencies [025a5a2]
+  - @crossmint/client-sdk-react-ui@1.16.0
+
 ## 0.1.77
 
 ### Patch Changes