+++ packages/wallets/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crossmint/wallets-sdk
 
+## 0.2.2
+
+### Patch Changes
+
+- ca39114: fixed codegen script
+
+## 0.2.1
+
+### Patch Changes
+
+- 9c71ce8: general bug fixes
+
 ## 0.2.0
 
 ### Minor Changes